Search dialog: add searching by part number/mfg code
The new "Components part number" checkbox is added to search dialog.
It enables searching the "mfgcode" string attached to each element.
It is dependent on existing Components checkbox.

For me it is useful for cases like answering
"Where did they put that P-mosfet in this revision of design?"
(board views I'm working with has a lot of mfgcode data - part number, Ohm values, even packages).